// src/types/beats/svm.ts
export interface SvmBlockTransactionsTransaction {
  signatures: string[];
  message: SvmBlockTransactionsTransactionMessage;
}
// ...
export interface SvmBlockTransactionsMeta {
  err: Record<string, unknown> | null;
  status: SvmBlockTransactionsMetaStatus;
  fee: number;
  preBalances: number[];
  postBalances: number[];
  innerInstructions: SvmBlockTransactionsMetaInnerinstructions[];
  logMessages: string[];
  preTokenBalances: SvmBlockTransactionsMetaPretokenbalances[];
  postTokenBalances: SvmBlockTransactionsMetaPosttokenbalances[];
  rewards: unknown[];
  loadedAddresses: SvmBlockTransactionsMetaLoadedaddresses;
  computeUnitsConsumed: number;
  costUnits?: number;
  returnData?: SvmBlockTransactionsMetaReturndata;
}

export interface SvmBlockTransactions {
  transaction: SvmBlockTransactionsTransaction;
  meta: SvmBlockTransactionsMeta;
  version: number | string;
}

export interface SvmBlockRewards {
  pubkey: string;
  lamports: number;
  postBalance: number;
  rewardType: string;
  commission: null;
}

export interface SvmBlock {
  previousBlockhash: string;
  blockhash: string;
  parentSlot: number;
  transactions: SvmBlockTransactions[];
  rewards: SvmBlockRewards[];
  blockTime: number;
  blockHeight: number;
  _network: string;
  _vm: string;
}